editoast: change the uic type from i64 to u32

+++ editoast/editoast_schemas/src/infra/operational_point.rs
@@ -80,7 +80,7 @@ pub struct OperationalPointSncfExtension {
 pub struct OperationalPointIdentifierExtension {
     #[schema(inline)]
     pub name: NonBlankString,
-    pub uic: i64,
+    pub uic: u32,
 }
 
 impl OSRDTyped for OperationalPoint {
